Improving typings of VirtualJoystick
