+++ front/src/Phaser/UserInput/UserInputManager.ts
@@ -1,8 +1,8 @@
-import type { Direction } from "../../types";
 import type { GameScene } from "../Game/GameScene";
 import { touchScreenManager } from "../../Touch/TouchScreenManager";
 import { MobileJoystick } from "../Components/MobileJoystick";
 import { enableUserInputsStore } from "../../Stores/UserInputStore";
+import type { Direction } from "phaser3-rex-plugins/plugins/virtualjoystick.js";
 
 interface UserInputManagerDatum {
     keyInstance: Phaser.Input.Keyboard.Key;
