+++ messages/package.json
@@ -36,7 +36,7 @@
   },
   "repository": {
     "type": "git",
-    "url": "https://github.com/CatsMiaow/node-grpc-typescript.git"
+    "url": "https://github.com/CatsMiaow/node-grpc-typescript.git#grpc_tools_node_protoc_ts"
   },
   "keywords": [
     "Node.js",
